functions/generate-password/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
qrcode[pil]==7.4.2
cryptography==41.0.7